irteqa-health-api/requirements.txt
# Add these to your existing requirements.txt

# Deepgram SDK
deepgram-sdk==3.2.7

# WebSocket support
websockets==12.0
python-multipart==0.0.6
python-dotenv==1.0.0

# For SOAP generation (AI)
openai==1.6.1  # or anthropic for Claude

# Existing dependencies (keep these)
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
aiosqlite==0.19.0
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
httpx==0.25.2
redis==5.0.1
celery==5.3.4